backend/requirements.txt
fastapi~=0.115.0
uvicorn[standard]~=0.34.0
anthropic~=0.52.0
google-genai~=1.0.0
pandas~=2.3.2
openpyxl~=3.1.5
pydantic~=2.11.0
requests~=2.32.0
pytest~=8.4.0
